refactor: 포맷팅, 폴더명 수정 (#10)
