+++ commitlint.config.mjs
@@ -12,18 +12,7 @@ export default {
     'type-enum': [
       2,
       'always',
-      [
-        'feat',
-        'fix',
-        'chore',
-        'style',
-        'hotfix',
-        'docs',
-        'refactor',
-        'test',
-        'init',
-        'build',
-      ],
+      ['feat', 'fix', 'chore', 'style', 'hotfix', 'docs', 'refactor', 'test', 'init', 'build'],
     ],
     'references-empty': [2, 'never'],
   },
