+++ src/pages/main/components/banner.tsx
@@ -1,0 +1,5 @@
+function Banner() {
+  return <div className="h-[20rem] bg-neutral-200" aria-label="프로모션 배너 자리" />;
+}
+
+export default Banner;
