--- src/pages/main_tmp/components/banner.tsx
@@ -1,7 +1,0 @@
-function Banner() {
-  return (
-    <div className="h-[20rem] bg-neutral-200" aria-label="프로모션 배너 자리" />
-  );
-}
-
-export default Banner;
