+++ src/pages/map/map-page.tsx
@@ -14,15 +14,9 @@ export default function MapPage() {
     <div className="h-dvh w-full">
       <MapDiv
         style={{ width: '100%', height: '100vh' }}
-        fallback={
-          <div className="grid h-full place-items-center">지도 로딩 중…</div>
-        }
+        fallback={<div className="grid h-full place-items-center">지도 로딩 중…</div>}
       >
-        <NaverMap
-          defaultCenter={fallback}
-          center={loc ?? fallback}
-          defaultZoom={16}
-        >
+        <NaverMap defaultCenter={fallback} center={loc ?? fallback} defaultZoom={16}>
           {loc && <Marker position={loc} />}
         </NaverMap>
       </MapDiv>
@@ -30,11 +24,7 @@ export default function MapPage() {
       {error && (
         <div className="absolute top-4 left-1/2 -translate-x-1/2 rounded bg-white/90 px-3 py-2 shadow">
           위치 권한/가져오기 실패: {error.message}
-          <button
-            className="ml-2 underline"
-            onClick={request}
-            disabled={loading}
-          >
+          <button className="ml-2 underline" onClick={request} disabled={loading}>
             다시 시도
           </button>
         </div>
