+++ src/shared/components/text-field/search-text-field.tsx
@@ -107,12 +107,7 @@ export default function SearchTextField({
           ].join(' ')}
         >
           {showSearchIcon && (
-            <Icon
-              name="search"
-              size={2}
-              ariaHidden
-              className="shrink-0 text-gray-400"
-            />
+            <Icon name="search" size={2} ariaHidden className="shrink-0 text-gray-400" />
           )}
 
           <input
@@ -133,17 +128,8 @@ export default function SearchTextField({
 
           {/* 클리어(X) — 값 있을 때만 */}
           {val?.length > 0 && (
-            <button
-              type="button"
-              onClick={handleClear}
-              aria-label="입력 지우기"
-            >
-              <Icon
-                className="text-gray-300"
-                name="clear"
-                size={2.4}
-                ariaHidden
-              />
+            <button type="button" onClick={handleClear} aria-label="입력 지우기">
+              <Icon className="text-gray-300" name="clear" size={2.4} ariaHidden />
             </button>
           )}
         </div>
