+++ src/shared/constants/icons.ts
@@ -1,7 +1,1 @@
-export const ICONS = [
-  "arrow",
-  "clear",
-  "home",
-  "search",
-  "x-icon"
-] as const;
+export const ICONS = ['arrow', 'clear', 'home', 'search', 'x-icon'] as const;
