+++ src/shared/mocks/products.ts
