+++ src/shared/routes/router.tsx
@@ -2,8 +2,8 @@ import { lazy } from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 
 const Layout = lazy(() => import('@/shared/layouts/layout.jsx'));
-const MainPage = lazy(() => import('@/pages/main_tmp/main-page.js'));
-const MapPage = lazy(() => import('@/pages/map_tmp/map-page'));
+const MainPage = lazy(() => import('@/pages/main/main-page.js'));
+const MapPage = lazy(() => import('@/pages/map/map-page'));
 
 export const router = createBrowserRouter([
   {
