+++ src/shared/types/product.ts
